Use status to determine icon type
