+++ ui/app/components/ui/item/item.component.js
@@ -1,18 +1,47 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
+import Icon from '../icon'
+
+const SendIcon = () => (
+  <Icon
+    type="send"
+    width={28}
+    height={28}
+    color="#2F80ED"
+  />
+)
+
+const InteractionIcon = () => (
+  <Icon
+    type="interaction"
+    width={28}
+    height={28}
+    color="#2F80ED"
+  />
+)
+
+const ApproveIcon = () => (
+  <Icon
+    type="approve"
+    width={28}
+    height={28}
+    color="#2F80ED"
+  />
+)
 
 const Item = ({
   className,
   status,
-  icon,
   title,
   subtitle,
   more,
   crypto,
   cash,
 }) => {
-  const active = status === 'approved'
+  const isApproved = status === 'approved'
+  const isPending = status === 'pending'
+
   if (status === 'unapproved') {
     subtitle = (
       <h3>
@@ -24,9 +53,15 @@ const Item = ({
   return (
     <div className={className}>
       <div className="col icon">
-        {icon}
+        {isApproved ? (
+          <ApproveIcon />
+        ) : isPending ? (
+          <InteractionIcon />
+        ) : (
+          <SendIcon />
+        )}
       </div>
-      <div className={`col main${active ? ' active' : ''}`}>
+      <div className={`col main${isApproved ? ' approved' : ''}`}>
         {typeof title === 'string' ? (
           <h2>{ title }</h2>
         ) : (
@@ -43,7 +78,7 @@ const Item = ({
           </div>
         )}
       </div>
-      <div className={`col amount${active ? ' active' : ''}`}>
+      <div className={`col amount${isApproved ? ' approved' : ''}`}>
         <h2>{crypto}</h2>
         <h3>{cash}</h3>
       </div>
@@ -52,15 +87,12 @@ const Item = ({
 }
 
 Item.defaultProps = {
-  active: false,
   status: 'pending',
 }
 
 Item.propTypes = {
   className: PropTypes.string,
-  active: PropTypes.bool,
   status: PropTypes.string,
-  icon: PropTypes.node,
   title: PropTypes.oneOfType([
     PropTypes.string,
     PropTypes.node,
