+++ ui/app/css/itcss/components/item.scss
@@ -90,7 +90,7 @@
   }
 
   .main, .amount {
-    &.active {
+    &.approved {
       h2 {
         color: #24292E;
       }
